Hopefully made function to check that current user is event host before modifying event will test later

+++ GatheringAPI/Services/DbGroupRepo.cs
@@ -171,16 +171,24 @@ namespace GatheringAPI.Services
             await _context.SaveChangesAsync();
         }
 
-        public async Task DeleteEventAsync(long groupId, long eventId)
+        public async Task DeleteEventAsync(long groupId, long eventId, long UserId)
         {
+            var @event = _context.Events.Find(eventId);
+
+            if (await HostMatchesCurrent(UserId, @event) == false)
+                throw new UnauthorizedAccessException();
+
             var groupEvent = await _context.GroupEvents.FindAsync(groupId, eventId);
 
             _context.GroupEvents.Remove(groupEvent);
             await _context.SaveChangesAsync();
         }
 
-        public async Task<bool> UpdateEventAsync(long groupId, Event @event)
+        public async Task<bool> UpdateEventAsync(long groupId, Event @event, long UserId)
         {
+            if (await HostMatchesCurrent(UserId, @event) == false)
+                throw new UnauthorizedAccessException();
+
             _context.Entry(@event).State = EntityState.Modified;
 
             try
@@ -317,6 +325,11 @@ namespace GatheringAPI.Services
             var user = await _context.Users.FirstOrDefaultAsync(u => u.UserName == userName);
             return user.Id;
         }
+
+        public bool HostMatchesCurrent(long current, Event @event)
+        {
+            return current == @event.EventHost.UserId;
+        }
     }
 
     public interface IGroup
@@ -324,7 +337,7 @@ namespace GatheringAPI.Services
         IEnumerable<GroupDto> GetAll(long userId);
 
         GroupDto Find(long id, long userId);
-        Task<bool> UpdateEventAsync(long groupId, Event @event);
+        Task<bool> UpdateEventAsync(long groupId, Event @event, long UserId);
         Task CreateAsync(Group group, long userId);
 
         Task<Group> DeleteAsync(long id, long userId);
@@ -333,13 +346,15 @@ namespace GatheringAPI.Services
 
         Task AddEventAsync(long groupId, long eventId);
 
-        Task DeleteEventAsync(long groupId, long eventId);
+        Task DeleteEventAsync(long groupId, long eventId, long UserId);
         Task AddUserAsync(long groupId, string userName);
         Task RemoveUserAsync(long groupId, long userId);
 
         void SendInvites(Event @event);
         Task CreateEventAsync(Event @event, long userId, long groupId);
         Task<long> FindUserIdByUserName(string userName);
+
+        bool HostMatchesCurrent(long current, Event @event);
     }
 
 }
